keeping log files in sn_cleanup

+++ SneakerNet.plugins/sn_cleanup.pl
@@ -15,7 +15,7 @@ use FindBin;
 use lib "$FindBin::RealBin/../lib/perl5";
 use SneakerNet qw/exitOnSomeSneakernetOptions recordProperties readConfig samplesheetInfo_tsv command logmsg fullPathToExec/;
 
-our $VERSION = "1.0";
+our $VERSION = "1.1";
 our $CITATION = "sn_cleanup.pl, by Lee Katz";
 
 # For any warnings in the SN report
@@ -36,6 +36,7 @@ sub main{
   die usage() if($$settings{help} || !@ARGV);
   $$settings{numcpus}||=1;
   $$settings{tempdir}||=File::Temp::tempdir(basename($0).".XXXXXX",TMPDIR=>1,CLEANUP=>1);
+  $$settings{debug} ||= 0; # need this value to be explicitly defined
 
   my $dir=$ARGV[0];
 
@@ -71,12 +72,14 @@ sub cleanup{
   my $numRemoved = 0;
   my $sizeRemoved= 0;
 
+  # Which files do we remove?
   my @rmFiles = uniq glob(
     "
       $dir/SneakerNet/assemblies/*/shovill
       $dir/SneakerNet/assemblies/*/prodigal
     "
   );
+
   for my $file(@rmFiles){
     my %removalStats = cleanThis($file, $$settings{debug}, $settings);
     $numRemoved += $removalStats{num};
@@ -94,16 +97,26 @@ sub cleanThis{
 
   if(-d $file){
     for my $f(glob("$file/*")){
+      # But keep files ending in .log
+      next if($f =~ /\.log$/);
+
       my %removalStats = cleanThis($f, $debug, $settings);
       $numRemoved += $removalStats{num};
       $sizeRemoved+= $removalStats{size};
     }
-    logmsg "RMDIR: $file";
-    $numRemoved++;
-    my $size = -s $file;
-    $sizeRemoved += $size;
-    if(!$$settings{debug}){
-      rmdir($file) or die "ERROR: could not rmdir $file: $!";
+
+    # Now all intended files have been removed.
+    # Remove the directory if it is empty.
+    if(is_folder_empty($file)){
+      logmsg "RMDIR: $file";
+      $numRemoved++;
+      my $size = -s $file;
+      $sizeRemoved += $size;
+      if(!$$settings{debug}){
+        rmdir($file) or die "ERROR: could not rmdir $file: $!";
+      }
+    } else {
+      logmsg "NOTE: $file/ not empty; not removing.";
     }
   }
   else {
@@ -118,6 +131,13 @@ sub cleanThis{
 
   #print Dumper {size=>$sizeRemoved, num=>$numRemoved};
   return (size=>$sizeRemoved, num=>$numRemoved);
+}
+
+# https://stackoverflow.com/a/4493532
+sub is_folder_empty {
+    my $dirname = shift;
+    opendir(my $dh, $dirname) or die "Not a directory";
+    return scalar(grep { $_ ne "." && $_ ne ".." } readdir($dh)) == 0;
 }
 
 sub usage{
